add tests for v1 and v2 manifests
